+++ src/Model/PhasorDynamics/SignalNode/SignalNodeImpl.hpp
@@ -52,7 +52,7 @@ namespace GridKit
     }
 
     template <class ScalarT, typename IdxT>
-    void SignalNode<ScalarT, IdxT>::set(ScalarT* signal, IdxT variable_index)
+    void SignalNode<ScalarT, IdxT>::set(ScalarT* signal, IdxT* variable_index)
     {
       signal_         = signal;
       variable_index_ = variable_index;